Add two new REST endpoints for cluster result /request and /cancel
Those two endpoints accept simple object {"id": <vmId>} that marks
a VM with the given id as required to run (or removes it from the
list of required VMs).

That causes the optimizer to compute a migration sequence that
ends up with the required VMs running in addition to being
"optimal" with regards to cluster policy.

Change-Id: I93b1474a6d66e3616f87a3145aa6a0cce28df302
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=877209
Signed-off-by: Martin Sivak <msivak@redhat.com>