webadmin: Fixed "Optimize start" button stuck.
Fixed the "Optimize start" button behaviour by making it disabled
when the virtual machine in focus is running. If multiple vms are
selected and at least one of them is 'down', the 'optimize start'
button will be enabled, otherwise, it will be false.

This corresponds with the patch in ovirt-engine:
http://gerrit.ovirt.org/#/c/32997/.

Change-Id: I942ba37047fe1ef572633c48d237d4dab26f7d20
Bug-Url: https://bugzilla.redhat.com/1140723
Signed-off-by: Tomer Saban <tsaban@redhat.com>