Do not propose starting a VM without explicit user's request
Probabilistic algorithms are funny and so there was a possibility
of a following case when balanced policy was selected:

Host1:
  - VM1

Host2:
  - empty

and VM2 in DOWN state.

The optimizer proposed to start the second VM to keep the cluster
in balanced state.

This fix penalizes VM starts that weren't requested by the user.

Change-Id: Ie7cf8ae8939454ab23ed488a6f771fab4228ae63