Stop bundling Optaplanner, download it instead
Optaplanner is expected to be installed (unpacked) on the filesystem
and the symlink /usr/share/java/ovirt-optimizer/optaplanner should
point to the binaries subdirectory of Optaplanner distribution.

A new helper script ovirt-optimizer-setup is included and can help
with this. It know how to download Optaplanner 6.2.0, verify its
integrity and unpack it to the right place.

Change-Id: I521dcf24fa12a0147076dd46fa950dbd8a4d7a85