Add debug endpoint
This adds two new HTTP endpoints:

GET /ovirt-optimizer/debug
POST /ovirt-optimizer/debug/<cluster id>

Use the GET endpoint to download the current fact database and
solutions and post the downloaded file back to the POST endpoint
ending with an existing cluster ID to compute the score again with
the uploaded data.

This allows us to test the rule engine by modifying the received
dump and testing whether the score improved or not. It is also
a way to get all needed data for filing a rule behaviour bug.

Both endpoints need to be enabled by setting the

org.ovirt.optimizer.debug property to true

The config file is consulted every time the endpoint is queried
so it is not necessary to restart the optimizer after enabling
or disabling the debug mode.

Change-Id: I6d9974dada109f7ec4f953a3633c67d5ac86ad86
Signed-off-by: Martin Sivak <msivak@redhat.com>