Add simple scheduling endpoint we can use as one-by-one external scheduler
This patch adds another debug endpoint:

POST /debug/<cluster>/<vm>

The post data are expected to be in the format returned by the
GET /debug url. The cluster and vm specify the cluster where
vm should start.

The result is a list of all available hosts with scores
representing the fittness of a solution where the VM starts
on the respective host.

When an existing solution is provided for the cluster then
the existing migration steps are considered as part of
the start sequence.

Change-Id: Idacddcd93a895945dc40cd98868e14ae573d7b39
Signed-off-by: Martin Sivak <msivak@redhat.com>